Move downloader to own file
